Revert "Merge reach wall hit check"
This reverts commit b5990080
